Refactor skinny-engine
- removing SkinnyEngineServletBase
- fixing possible NPE only when testing
- renaming AsyncSupport to AsyncRoutingDsl
- renaming CoreDsl to CoreRoutingDsl
