+++ engine/src/main/scala/skinny/engine/routing/Route.scala
@@ -53,7 +53,7 @@ object Route {
   def apply(transformers: Seq[RouteTransformer], action: Action): Route =
     apply(transformers, action, (_: HttpServletRequest) => "")
 
-  // TODO: remove HttpServletRequest
+  // TODO: remove HttpServletRequest from contextPath
   def apply(transformers: Seq[RouteTransformer], action: Action, contextPath: HttpServletRequest => String): Route = {
     val route = Route(action = action, contextPath = contextPath)
     transformers.foldLeft(route) { (route, transformer) => transformer(route) }
